+++ src/modules/person-kontext/domain/personenkontext.service.ts
@@ -4,8 +4,8 @@ import { DomainError, EntityCouldNotBeDeleted, EntityNotFoundError } from '../..
 import { Paged } from '../../../shared/paging/paged.js';
 import { ScopeOrder } from '../../../shared/persistence/scope.enums.js';
 import { PersonRepo } from '../../person/persistence/person.repo.js';
-import { PersonenkontextRepo } from '../../person/persistence/personenkontext.repo.js';
-import { PersonenkontextScope } from '../../person/persistence/personenkontext.scope.js';
+import { PersonenkontextRepo } from '../persistence/personenkontext.repo.js';
+import { PersonenkontextScope } from '../persistence/personenkontext.scope.js';
 import { PersonDo } from '../../person/domain/person.do.js';
 import { PersonenkontextDo } from './personenkontext.do.js';
 import { MismatchedRevisionError } from '../../../shared/error/mismatched-revision.error.js';
