+++ src/modules/person-kontext/persistence/personenkontext.scope.integration-spec.ts
@@ -11,8 +11,8 @@ import {
     MapperTestModule,
 } from '../../../../test/utils/index.js';
 import { ScopeOrder } from '../../../shared/persistence/scope.enums.js';
-import { PersonenkontextDo } from '../../person-kontext/domain/personenkontext.do.js';
-import { PersonPersistenceMapperProfile } from './person-persistence.mapper.profile.js';
+import { PersonenkontextDo } from '../domain/personenkontext.do.js';
+import { PersonPersistenceMapperProfile } from '../../person/persistence/person-persistence.mapper.profile.js';
 import { PersonenkontextEntity } from './personenkontext.entity.js';
 import { PersonenkontextScope } from './personenkontext.scope.js';
