+++ src/modules/person/persistence/person-persistence.mapper.profile.spec.ts
@@ -7,7 +7,7 @@ import { PersonDo } from '../domain/person.do.js';
 import { PersonEntity } from '../persistence/person.entity.js';
 import { PersonPersistenceMapperProfile } from './person-persistence.mapper.profile.js';
 import { PersonenkontextDo } from '../../person-kontext/domain/personenkontext.do.js';
-import { PersonenkontextEntity } from './personenkontext.entity.js';
+import { PersonenkontextEntity } from '../../person-kontext/persistence/personenkontext.entity.js';
 
 describe('PersonPersistenceMapperProfile', () => {
     let module: TestingModule;
